requirements.txt
contractions~=0.1.72
pyenchant~=3.2.2
fitz~=0.0.1.dev2
nltk~=3.7
XlsxWriter~=3.0.3
spellchecker~=0.4
pdf2image~=1.16.0
pytesseract~=0.3.10
streamlit~=1.13.0
pandas~=1.5.0
numpy~=1.23.3
Pillow~=9.2.0
openai~=0.23.1
fpdf~=1.7.2
matplotlib~=3.6.0
sklearn~=0.0
scikit-learn~=1.1.2
wordcloud~=1.8.2.2
yellowbrick~=1.5